asobi6
sanitize関数を自作し、安全対策関数として利用。before変数は呼び出し元のプログラムから$_POSTを受け取っている。after変数には安全対策されたデータの結果を入れている。
